// src/app/models/Product.ts
export interface Product {
    productId: string;
    productName: string;
    prodCategory: string;
    prodSubCategory: string;
    prodPrice: number;
    discountPrice: number;
    quantity: number;
    prodAvailability: string;
    productDesc: string;
    imageUrl: string;
    videoUrl: string;
    brand: string;
    productSize: string;
    rating: number;
    sellerId: number;
    totalReviews: number;
    creationTime: string;
    lastUpdateTime: string;
  }
  